Implement walking of syntax nodes and semantic analysis of types
To do: Add unit and integration tests in addition to existing system tests, and ensure that the type symbols and declaration symbols are only extracted from the semantic model when they are needed. Also, already start considering what a performance profile might look like for this architecture. I'm a little wary of the amount of heap allocations being performed, especially on larger chunks of code being analyzed.